Accidentally overwrote my changes to Gaussian elimination, but I wasn't able to achieve any performance gain through them anyway.  I have at least sped up partial pivoting to be less than half a second, which was my goal for partial pivoting.
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@4526 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ examples/hplpx/hplmatrex/server/LUblock.hpp
@@ -38,7 +38,7 @@ LUblock::LUblock(unsigned int h, unsigned int w){
 //    columns = ((int*) workSpace) + 1;
     rows = h;
     columns = w;
-    for(int i = 0;i < h; i++){
+    for(int i = 0;i < (int)h; i++){
         data[i] = workSpace + 8 + i*w;
     }
 }
